refactor method to helper
